fix getWeekDay & move db.js to admin-settings.js
