+++ src/websocket-listeners/SC-module/lib/db-settings-backup.js
@@ -2,7 +2,6 @@ const jsonFiles               = require('./jsonFiles');
 const path 										= require('path')
 const fs 											= require('fs');
 let filepath = "db-settings=FilePath";
-//const filepath = './SC-module/db/db-settings.json';
 
 async function getFilepath(){
   var myPath = "";
@@ -11,29 +10,6 @@ async function getFilepath(){
       var self = this;
       var filname = "db-settings"
       myPath = path.join(__dirname,"db", filname + ".json");
-      //
-      // if (fs.existsSync(myPath)) {
-      //   console.log("scheduleSettings_path file exist");
-      //   });
-      //
-      //
-      // } else {
-      //   console.log("scheduleSettings_path does not exist");
-      //   // var data = {
-      //   //   "schedule": [{
-      //   //     title: "First row added",
-      //   //     startTime: "12:00",
-      //   //     cueLength: "00:01:00",
-      //   //     cueBool: false,
-      //   //     fiveBool: false
-      //   //   }]
-      //   // };
-      //   //
-      //   // fs.writeFile(scheduleSettings_path, JSON.stringify(data, null, 4), function(err) {
-      //   //   if (err) throw err;
-      //   //   console.log('Saved!');
-      //   // });
-      // }
     } catch (err) {
       console.log(err);
     }
